requirements.txt
numpy==1.24.1
Pillow==9.4.0
streamlit==1.19.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
